rename test package; markdown syntax fix; jack of all trades
